Extract common logic to setPrivacyScreen helper
