+++ android/app/src/main/java/org/oasisprotocol/wallet/MainActivity.java
@@ -10,6 +10,7 @@ public class MainActivity extends BridgeActivity {
         super.onCreate(savedInstanceState);
 
         // Set secure flag to prevent app content from being captured in screenshots or recent apps view
+        // Should be control by @capacitor plugin in the app, keeping this for safety
         getWindow().setFlags(WindowManager.LayoutParams.FLAG_SECURE,
                              WindowManager.LayoutParams.FLAG_SECURE);
     }
